Move discover to its own module
