+++ soco/alarms.py
@@ -8,7 +8,8 @@ import logging
 from datetime import datetime
 import re
 import weakref
-from .core import discover, PLAY_MODES
+from .core import PLAY_MODES
+from .discovery import discover
 from .xml import XML
 
 log = logging.getLogger(__name__)  # pylint: disable=C0103
